requirements.txt
environs
asgiref==3.2.7
Django==3.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
drf-yasg==1.17.1
sendgrid==6.4.1
certifi
django-crontab==0.7.1
boto3==1.14.20
django-storages==1.9.1
uuid
